feat: complet project repo

+++ pkg/usecase/usecase/project.go
@@ -2,24 +2,21 @@ package usecase
 
 import (
 	"github.com/athmos-cloud/infra-worker-athmos/pkg/adapter/context"
+	"github.com/athmos-cloud/infra-worker-athmos/pkg/adapter/dto"
 	arepo "github.com/athmos-cloud/infra-worker-athmos/pkg/adapter/repository"
 	"github.com/athmos-cloud/infra-worker-athmos/pkg/domain/model"
 	"github.com/athmos-cloud/infra-worker-athmos/pkg/kernel/errors"
 	"github.com/athmos-cloud/infra-worker-athmos/pkg/kernel/option"
+	"github.com/athmos-cloud/infra-worker-athmos/pkg/kernel/share"
 	"github.com/athmos-cloud/infra-worker-athmos/pkg/usecase/repository"
 )
 
-const (
-	ProjectIDKeyContext    = "project_id"
-	ProjectOwnerKeyContext = "project_owner"
-)
-
 type Project interface {
-	Get(context.Context, *model.Project) *model.Project
-	List(context.Context, []*model.Project) []*model.Project
-	Create(context.Context, chan *model.Project, chan errors.Error)
-	Update(context.Context, *model.Project) *model.Project
-	Delete(context.Context, *model.Project)
+	Get(context.Context, *model.Project) errors.Error
+	List(context.Context, *[]model.Project) errors.Error
+	Create(context.Context, *model.Project) errors.Error
+	Update(context.Context, *model.Project) errors.Error
+	Delete(context.Context, *model.Project) errors.Error
 }
 
 func NewProjectUseCase(repo repository.Project) Project {
@@ -30,31 +27,46 @@ type projectUseCase struct {
 	repo repository.Project
 }
 
-func (pu *projectUseCase) Get(ctx context.Context, project *model.Project) *model.Project {
-	return pu.repo.Find(ctx, option.Option{
+func (pu *projectUseCase) Get(ctx context.Context, project *model.Project) errors.Error {
+	foundProject, err := pu.repo.Find(ctx, option.Option{
 		Value: arepo.FindByIDRequest{
-			ID: ctx.Value(ProjectIDKeyContext).(string),
+			ID: ctx.Value(share.ProjectIDKey).(string),
 		},
 	})
+	if !err.IsOk() {
+		return err
+	}
+	*project = *foundProject
+	return errors.OK
 }
 
-func (pu *projectUseCase) List(ctx context.Context, project []*model.Project) []*model.Project {
-	project = pu.repo.FindAll(ctx, option.Option{
+func (pu *projectUseCase) List(ctx context.Context, projects *[]model.Project) errors.Error {
+	foundProjects, err := pu.repo.FindAll(ctx, option.Option{
 		Value: arepo.FindAllByOwnerRequest{
-			Owner: ctx.Value(ProjectOwnerKeyContext).(string),
+			Owner: ctx.Value(share.OwnerIDKey).(string),
 		},
 	})
-	return project
+	*projects = *foundProjects
+	return err
 }
 
-func (pu *projectUseCase) Create(ctx context.Context, projectCh chan *model.Project, errCh chan errors.Error) {
-	pu.repo.Create(ctx, projectCh, errCh)
+func (pu *projectUseCase) Create(ctx context.Context, project *model.Project) errors.Error {
+	return pu.repo.Create(ctx, project)
 }
 
-func (pu *projectUseCase) Update(ctx context.Context, project *model.Project) *model.Project {
+func (pu *projectUseCase) Update(ctx context.Context, project *model.Project) errors.Error {
+	project, err := pu.repo.Find(ctx, option.Option{Value: arepo.FindByIDRequest{ID: ctx.Value(share.ProjectIDKey).(string)}})
+	if !err.IsOk() {
+		return err
+	}
+	project.Name = ctx.Value(share.RequestContextKey).(dto.UpdateProjectRequest).Name
 	return pu.repo.Update(ctx, project)
 }
 
-func (pu *projectUseCase) Delete(ctx context.Context, project *model.Project) {
-	pu.repo.Delete(ctx, project)
+func (pu *projectUseCase) Delete(ctx context.Context, project *model.Project) errors.Error {
+	project, err := pu.repo.Find(ctx, option.Option{Value: arepo.FindByIDRequest{ID: ctx.Value(share.ProjectIDKey).(string)}})
+	if !err.IsOk() {
+		return err
+	}
+	return pu.repo.Delete(ctx, project)
 }
